add grid component
